don't send empty GameInf packets and resend them on reconnect
Resend because: on server restarts the server will lose the stage information.

If only one client is connected to the server, the GameInf packet currently isn't resent, so the server doesn't know in which stage the client is
(which I'd like to display on the website).

With more then one client connected it already works, because when another client joins the server, the client will send its GameInf packet.